+++ src/routes/+layout.server.ts
@@ -1,6 +1,5 @@
 import type { PartialDocsConfig } from '@keenmate/svelte-docs';
 import { mergeConfig, defaultConfig, generateSSRStyles } from '@keenmate/svelte-docs';
-import pkg from '@keenmate/web-daterangepicker/package.json';
 
 // Enable prerendering for all routes
 export const prerender = true;
@@ -76,7 +75,6 @@ export async function load() {
 
 	return {
 		config: mergedConfig,
-		ssrStyles,
-		version: pkg.version
+		ssrStyles
 	};
 }